+++ src/Client.js
@@ -33,6 +33,19 @@ export default class Client {
 			});
 	}
 
+	/**
+	 * Get list of library versions
+	 * @param  {String} name Name of the library to fetch
+	 * @param  {Object} query The query parameters for versions. See Particle.getLibraryVersions
+	 * @return {Promise}
+	 */
+	libraryVersions(name, query = {}) {
+		return this.api.getLibraryVersions(Object.assign({}, query, { name, auth: this.auth }))
+			.then(payload => {
+				const libraries = payload.body.data || [];
+				return libraries.map(l => new Library(this, l));
+			});
+	}
 
 	/**
 	 * Publish a new library version
@@ -45,19 +58,33 @@ export default class Client {
 				const library = payload.body.data || {};
 				return new Library(this, library);
 			}, error => {
-				if (error.body && error.body.errors) {
-					const errorMessages = error.body.errors.map((e) => e.message).join('\n');
-					throw new Error(errorMessages);
-				}
-				throw error;
-
+				this._throwError(error);
 			});
-		// TODO: format error
 	}
 
 
+	/**
+	 * Delete an entire published library
+	 * @param  {String} $0.name Name of the library to delete
+	 * @param  {String} $0.force Key to force deleting a public library
+	 * @return {Promise}
+	 */
+	deleteLibrary({ name, version, force }) {
+		return this.api.deleteLibrary({ name, force, auth: this.auth })
+			.then(payload => {
+				return true;
+			}, error => {
+				this._throwError(error);
+			});
+	}
 
-
+	_throwError(error) {
+		if (error.body && error.body.errors) {
+			const errorMessages = error.body.errors.map((e) => e.message).join('\n');
+			throw new Error(errorMessages);
+		}
+		throw error;
+	}
 
 	downloadFile(url) {
 		return this.api.downloadFile({ url });